% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{build_heatmap_data}
\alias{build_heatmap_data}
\title{Build one comparison's pathway x kinase LogFC matrix}
\usage{
build_heatmap_data(pathways_df, nodes_df, min_kinases)
}
\description{
Build one comparison's pathway x kinase LogFC matrix
}
\keyword{internal}
