% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichr_query}
\alias{enrichr_query}
\title{Thin wrapper over \code{enrichR::enrichr()} returning one long data frame}
\usage{
enrichr_query(genes, databases)
}
\description{
Thin wrapper over \code{enrichR::enrichr()} returning one long data frame
}
\keyword{internal}
