% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{join_pathway_column}
\alias{join_pathway_column}
\title{Join a \code{pathway} column (comma-joined pathway names per protein) onto \code{nodes}}
\usage{
join_pathway_column(nodes, short)
}
\description{
Join a \code{pathway} column (comma-joined pathway names per protein) onto \code{nodes}
}
\keyword{internal}
